New. Handle basic authorization under Test Subjects resource.

+++ app/Http/Controllers/Admin/Tests/SubjectsController.php
@@ -9,18 +9,34 @@ use App\Models\TestSubject;
 
 class SubjectsController extends AdminController
 {
-    public function showNewSubjectForm()
-    {
-        return view('pages.admin.subjects-new');
-    }
-
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showAll()
     {
+        $this->authorize('view-subjects');
+
         return view('pages.admin.subjects-list', [
             'subjects' => TestSubject::withCount('tests')->get()
         ]);
     }
 
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
+    public function showNewSubjectForm()
+    {
+        $this->authorize('create-subjects');
+
+        return view('pages.admin.subjects-new');
+    }
+
+    /**
+     * @param CreateSubjectRequest $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function newSubject(CreateSubjectRequest $request)
     {
         $validated = $request->validated();
@@ -31,8 +47,14 @@ class SubjectsController extends AdminController
         ]);
     }
 
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showSingleSubject()
     {
+        $this->authorize('view-subjects');
+
         $subject = $this->urlManager->getCurrentSubject();
         $subject->tests->loadCount('nativeQuestions as questions_count');
 
@@ -41,13 +63,23 @@ class SubjectsController extends AdminController
         ]);
     }
 
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showUpdateSubjectForm()
     {
+        $this->authorize('update-subjects');
+
         return view('pages.admin.subjects-single-settings', [
             'subject' => $this->urlManager->getCurrentSubject()
         ]);
     }
 
+    /**
+     * @param UpdateSubjectRequest $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function updateSubject(UpdateSubjectRequest $request)
     {
         $subject = $this->urlManager->getCurrentSubject();
@@ -58,8 +90,14 @@ class SubjectsController extends AdminController
         ]);
     }
 
+    /**
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function deleteSubject()
     {
+        $this->authorize('delete-subjects');
+
         $subject = $this->urlManager->getCurrentSubject();
         $subject->delete();
